黑白棋修正吃子逻辑

+++ reversi/src/main/java/bean/Move.java
@@ -14,6 +14,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Move {
-	private int row;
-	private int col;
+	private byte row;
+	private byte col;
 }
